# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisRun": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Analysis Id",
            "type": "string"
          },
          "candidate_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Candidate Id",
            "type": "string"
          },
          "corpus_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Snapshot"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_count": {
            "minimum": 0.0,
            "title": "Result Count",
            "type": "integer"
          },
          "resume_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Resume Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "queued",
              "extracting",
              "profiling",
              "matching",
              "ready",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "analysis_id",
          "candidate_id",
          "resume_id",
          "state",
          "created_at",
          "updated_at",
          "corpus_snapshot",
          "warnings",
          "error",
          "result_count"
        ],
        "title": "AnalysisRun",
        "type": "object"
      },
      "Body_upload_resume_api_v1_resumes_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_v1_resumes_post",
        "type": "object"
      },
      "CandidateProfile": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Candidate Id",
            "type": "string"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/Education"
            },
            "title": "Education",
            "type": "array"
          },
          "estimated_experience_years": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Experience Years"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence",
            "type": "array"
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/Experience"
            },
            "title": "Experience",
            "type": "array"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "extraction_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Extraction Warnings",
            "type": "array"
          },
          "job_titles": {
            "items": {
              "type": "string"
            },
            "title": "Job Titles",
            "type": "array"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "matching_enabled": {
            "title": "Matching Enabled",
            "type": "boolean"
          },
          "organizations": {
            "items": {
              "type": "string"
            },
            "title": "Organizations",
            "type": "array"
          },
          "profile_version": {
            "minimum": 1.0,
            "title": "Profile Version",
            "type": "integer"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "title": "Projects",
            "type": "array"
          },
          "resume_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Resume Id",
            "type": "string"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillEvidence"
            },
            "title": "Skills",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "candidate_id",
          "resume_id",
          "profile_version",
          "language",
          "summary",
          "skills",
          "education",
          "job_titles",
          "organizations",
          "experience",
          "estimated_experience_years",
          "projects",
          "evidence",
          "extraction_warnings",
          "matching_enabled",
          "expires_at"
        ],
        "title": "CandidateProfile",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvidenceRef": {
        "additionalProperties": false,
        "properties": {
          "chunk_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Document Id",
            "type": "string"
          },
          "document_version": {
            "minimum": 1.0,
            "title": "Document Version",
            "type": "integer"
          },
          "end": {
            "exclusiveMinimum": 0.0,
            "title": "End",
            "type": "integer"
          },
          "excerpt": {
            "minLength": 1,
            "title": "Excerpt",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "start": {
            "minimum": 0.0,
            "title": "Start",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "document_version",
          "chunk_id",
          "start",
          "end",
          "excerpt"
        ],
        "title": "EvidenceRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobPosting": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "company": {
            "minLength": 1,
            "title": "Company",
            "type": "string"
          },
          "content_version": {
            "minimum": 1.0,
            "title": "Content Version",
            "type": "integer"
          },
          "data_origin": {
            "enum": [
              "fixture",
              "permitted_source"
            ],
            "title": "Data Origin",
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence",
            "type": "array"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "job_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Job Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "other_requirements": {
            "items": {
              "type": "string"
            },
            "title": "Other Requirements",
            "type": "array"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/JobRequirement"
            },
            "title": "Requirements",
            "type": "array"
          },
          "source_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Source Id",
            "type": "string"
          },
          "source_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "source_id",
          "source_url",
          "content_version",
          "title",
          "company",
          "description",
          "summary",
          "requirements",
          "other_requirements",
          "evidence",
          "language",
          "fetched_at",
          "last_seen_at",
          "active",
          "data_origin"
        ],
        "title": "JobPosting",
        "type": "object"
      },
...
      "MatchExplanation": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Candidate Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence",
            "type": "array"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "job_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Job Id",
            "type": "string"
          },
          "job_version": {
            "minimum": 1.0,
            "title": "Job Version",
            "type": "integer"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "profile_version": {
            "minimum": 1.0,
            "title": "Profile Version",
            "type": "integer"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "scoring_version": {
            "minLength": 1,
            "title": "Scoring Version",
            "type": "string"
          },
          "state": {
            "enum": [
              "pending",
              "ready",
              "unavailable"
            ],
            "title": "State",
            "type": "string"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "candidate_id",
          "revision",
          "profile_version",
          "job_id",
          "job_version",
          "scoring_version",
          "state",
          "text",
          "strengths",
          "gaps",
          "evidence",
          "model_version",
          "prompt_version",
          "generated_at"
        ],
        "title": "MatchExplanation",
        "type": "object"
      },
...
      "RecommendationSet": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Candidate Id",
            "type": "string"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "profile_version": {
            "minimum": 1.0,
            "title": "Profile Version",
            "type": "integer"
          },
          "refresh_state": {
            "enum": [
              "idle",
              "pending",
              "failed"
            ],
            "title": "Refresh State",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "title": "Results",
            "type": "array"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "profile_version",
          "revision",
          "updated_at",
          "refresh_state",
          "results"
        ],
        "title": "RecommendationSet",
        "type": "object"
      },
...
      "UploadAccepted": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Analysis Id",
            "type": "string"
          },
          "candidate_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Candidate Id",
            "type": "string"
          },
          "resume_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Resume Id",
            "type": "string"
          },
          "status_url": {
            "minLength": 1,
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "resume_id",
          "analysis_id",
          "status_url"
        ],
        "title": "UploadAccepted",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Shared bootstrap. /api/v1 handlers return 501 until their owners implement them. /dev/fixtures returns synthetic data only; production mode is disabled.",
    "title": "AI Career Match API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "operationId": "analysis_status_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Planned feature; not implemented."
          }
        },
        "summary": "Analysis Status",
        "tags": [
          "planned"
        ],
        "x-implementation-status": "planned"
      }
    },
    "/api/v1/candidates/{candidate_id}/recommendations": {
      "get": {
        "operationId": "candidate_recommendations_api_v1_candidates__candidate_id__recommendations_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Planned feature; not implemented."
          }
        },
        "summary": "Candidate Recommendations",
        "tags": [
          "planned"
        ],
        "x-implementation-status": "planned"
      }
    },
    "/api/v1/candidates/{candidate_id}/recommendations/{revision}/jobs/{job_id}/explanation": {
      "get": {
        "description": "Read a generated explanation without triggering generation.",
        "operationId": "explanation_status_api_v1_candidates__candidate_id__recommendations__revision__jobs__job_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchExplanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not this session's candidate."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown revision or job."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explanation Status",
        "tags": [
          "explanations"
        ],
        "x-implementation-status": "planned"
      },
      "post": {
        "description": "Generate (or return the cached) explanation for one ranked job.",
        "operationId": "request_explanation_api_v1_candidates__candidate_id__recommendations__revision__jobs__job_id__explanation_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchExplanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not this session's candidate."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown revision or job."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Explanation",
        "tags": [
          "explanations"
        ],
        "x-implementation-status": "planned"
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "job_detail_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPosting"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Planned feature; not implemented."
          }
        },
        "summary": "Job Detail",
        "tags": [
          "planned"
        ],
        "x-implementation-status": "planned"
      }
    },
    "/api/v1/resumes": {
      "post": {
        "operationId": "upload_resume_api_v1_resumes_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_v1_resumes_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Planned feature; not implemented."
          }
        },
        "summary": "Upload Resume",
        "tags": [
          "planned"
        ],
        "x-implementation-status": "planned"
      }
    },
    "/api/v1/resumes/{resume_id}/profile": {
      "get": {
        "operationId": "candidate_profile_api_v1_resumes__resume_id__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Planned feature; not implemented."
          }
        },
        "summary": "Candidate Profile",
        "tags": [
          "planned"
        ],
        "x-implementation-status": "planned"
      }
    },
...
  }
}